add API Post

+++ src/controller/vacation.js
@@ -1,6 +1,9 @@
 import _throw from '#root/utils/_throw';
 import asyncWrapper from '#root/middleware/asyncWrapper';
 import Vacations from '#root/model/vacations';
+import pipelineLookup from '#root/config/pipelineLookup';
+import checkForbidden from '#root/utils/checkForbidden';
+import checkAuthor from '#root/utils/checkAuthor';
 import mongoose from 'mongoose';
 
 const vacationController = {
@@ -23,45 +26,63 @@ const vacationController = {
       { $setWindowFields: { output: { total: { $count: {} } } } },
 
       //Add 2 new fields with field page is current page user wanna get, and field pages is total pages divided by length of array and itemOfPage
-      {
-        $addFields: {
-          page: validPage,
-          pages: { $ceil: { $divide: ['$total', itemOfPage] } },
-        },
-      },
+      { $addFields: { page: validPage, pages: { $ceil: { $divide: ['$total', itemOfPage] } } } },
 
-      //Remove some firstN element in array if page >1
+      //Remove some firstN element in array if page > 1 and get quantity of element equal to itemOfPage
       { $skip: (validPage - 1) * itemOfPage },
-
-      //   //Only get quantity of element equal to itemOfPage
       { $limit: itemOfPage },
 
-      //Get images and videos by lookup to posts model by _id
+      //Get username of author by lookup to users model by userId
+      ...pipelineLookup.getUsername,
+
       {
         $lookup: {
           from: 'posts',
-          pipeline: [
-            { $unwind: '$subAlbum' },
-            { $sample: { size: 3 } },
-            { $group: { _id: '$subAlbum', subAlbum: { $push: '$subAlbum' } } },
-            { $unset: '_id' },
-          ],
           localField: '_id',
           foreignField: 'vacationId',
-          as: 'subAlbum',
-        },
-      },
+          pipeline: [
+            //Get total like by looking up to likes model
+            {
+              $lookup: {
+                from: 'likes',
+                pipeline: [{ $count: 'total' }],
+                localField: '_id',
+                foreignField: 'postId',
+                as: 'likes',
+              },
+            },
+            { $unwind: '$likes' },
+            { $addFields: { likes: '$likes.total' } },
 
-      //Get username of author by lookup to users model by userId
-      {
-        $lookup: {
-          from: 'users',
-          localField: 'userId',
-          foreignField: '_id',
-          pipeline: [{ $project: { username: 1, _id: 0 } }],
-          as: 'username',
+            //Get total comment by looking up to comment model
+            {
+              $lookup: {
+                from: 'comments',
+                pipeline: [{ $count: 'total' }],
+                localField: '_id',
+                foreignField: 'postId',
+                as: 'comments',
+              },
+            },
+            { $unwind: '$comments' },
+            { $addFields: { comments: '$comments.total' } },
+
+            //Only get field views, totalLikes, totalComments
+            {
+              $group: {
+                _id: '$vacationId',
+                views: { $sum: '$views' },
+                likes: { $sum: '$likes' },
+                comments: { $sum: '$comments' },
+              },
+            },
+            { $unset: '_id' },
+          ],
+          as: 'posts',
         },
       },
+      { $unwind: '$posts' },
+      { $addFields: { views: '$posts.views', likes: '$posts.likes', comments: '$posts.comments' } },
 
       //Set up new array with total field is length of array and list field is array without __v field
       {
@@ -77,23 +98,7 @@ const vacationController = {
             },
             { $project: { _id: 0 } },
           ],
-          data: [
-            {
-              $project: {
-                title: 1,
-                shareStatus: 1,
-                startingTime: 1,
-                endingTIme: 1,
-                lastUpdateAt: 1,
-                subAlbum: 1,
-                // userId: 1,
-                username: 1,
-              },
-            },
-            { $unwind: '$username' },
-            { $unwind: '$subAlbum' },
-            { $addFields: { username: '$username.username', subAlbum: '$subAlbum.subAlbum' } },
-          ],
+          data: [{ $project: { total: 0, page: 0, pages: 0, userId: 0, __v: 0, posts: 0 } }],
         },
       },
 
@@ -105,89 +110,25 @@ const vacationController = {
   }),
 
   getOne: asyncWrapper(async (req, res) => {
-    //Get id from params
     const { id } = req.params;
-
-    //Find vacation based on id
-    const foundVacation = await Vacations.findById(id);
-
-    //Throw an error if cannot find vacation
-    !foundVacation &&
-      _throw({ code: 400, errors: [{ field: 'id', message: 'invalid' }], message: 'vacation did not exist' });
-
-    //Get userId from verifyJWT middleware
     const userId = req.userInfo._id.toString();
-    switch (foundVacation.shareStatus) {
-      case 'protected':
-        //Throw an error if user is not in shareList of vacation
-        !foundVacation.shareList.includes(userId) &&
-          _throw({
-            code: 403,
-            errors: [{ field: 'shareList', message: 'user is not in shareList of this vacation' }],
-            message: 'Forbidden',
-          });
-        break;
 
-      case 'onlyme':
-        //Throw an error if shareStatus of vacation is onlyme and user is not author
-        foundVacation.userId !== userId &&
-          _throw({
-            code: 403,
-            errors: [{ field: 'shareStatus', message: 'this vacation only can be seen by its author' }],
-            message: 'Forbidden',
-          });
-        break;
-
-      default:
-        break;
-    }
+    //Throw an error if user login is not in shareList
+    await checkForbidden(userId, id);
 
     const result = await Vacations.aggregate([
       { $match: { _id: new mongoose.Types.ObjectId(id) } },
       {
         $lookup: {
-          from: 'posts',
-          pipeline: [
-            { $project: { vacationId: 0, __v: 0 } },
-            {
-              $lookup: {
-                from: 'users',
-                localField: 'userId',
-                foreignField: '_id',
-                pipeline: [{ $project: { username: 1, _id: 0 } }],
-                as: 'username',
-              },
-            },
-            { $unwind: '$username' },
-            { $addFields: { username: '$username.username' } },
-            {
-              $lookup: {
-                from: 'locations',
-                localField: 'locationId',
-                foreignField: '_id',
-                pipeline: [{ $project: { title: 1, _id: 0 } }],
-                as: 'location',
-              },
-            },
-            { $unwind: '$location' },
-            { $addFields: { location: '$location.title' } },
-          ],
-          localField: '_id',
-          foreignField: 'vacationId',
-          as: 'posts',
-        },
-      },
-      {
-        $lookup: {
           from: 'users',
           localField: 'userId',
           foreignField: '_id',
-          pipeline: [{ $project: { username: 1, _id: 0 } }],
-          as: 'username',
+          pipeline: [{ $project: { username: 1, avatar: 1 } }],
+          as: 'userInfo',
         },
       },
-      { $unwind: '$username' },
-      { $addFields: { username: '$username.username' } },
+      { $project: { userId: 0, __v: 0 } },
+      { $unwind: '$userInfo' },
     ]);
 
     //Send to front
@@ -230,44 +171,39 @@ const vacationController = {
   update: asyncWrapper(async (req, res) => {
     const { id } = req.params;
 
-    //Throw an error if cannot find Vacation
-    const foundVacation = await Vacations.findById(id);
-    !foundVacation &&
-      _throw({ code: 400, errors: [{ field: 'id', message: 'vacationId did not existed' }], message: 'invalid id' });
-
-    //Throw an error if user is not author of vacation
-    const userId = req.userInfo._id.toString();
-    foundVacation.userId.toString() !== userId &&
-      _throw({
-        code: 400,
-        errors: [{ field: 'userId', message: 'userId is not author of this vacation' }],
-        message: 'invalid userId',
-      });
-
-    //Get vital information from req.body
-    const { memberList, shareStatus, shareList } = req.body;
-    const updateKeys = ['title', 'description', 'memberList', 'shareStatus', 'shareList', 'startingTime', 'endingTime'];
-
-    //if memberList receive is not an array, then return memberlist contain only userId, otherwises, combine memberList and userId
-    const newMemberList = Array.isArray(memberList) ? [...new Set([...memberList, userId])] : [userId];
-
-    //If shareStatus is protected, and shareList is an array, then return combination of newMemberList and shareList, otherwise, return newMemberList, if shareStatus is not protected, then return null
-    const newShareList =
-      shareStatus === 'protected'
-        ? Array.isArray(shareList)
-          ? [...new Set(newMemberList.concat(shareList))]
-          : newMemberList
-        : null;
+    //Check whether user login is author of this vacation, if user is not author, then throw an error Forbidden
+    const foundVacation = await checkAuthor({ type: 'vacation', vacationId: id, userId: req.userInfo._id });
 
     //Save new info to foundVacation
+    const { memberList, shareStatus, shareList } = req.body;
+    const updateKeys = ['title', 'description', 'memberList', 'shareStatus', 'shareList', 'startingTime', 'endingTime'];
     updateKeys.forEach(key => {
       switch (key) {
         case 'memberList':
+          //if memberList receive is not an array, then return memberlist contain only userId, otherwises, combine memberList and userId
+          const newMemberList = Array.isArray(memberList) ? [...new Set([...memberList, userId])] : [userId];
           foundVacation.memberList = newMemberList;
           break;
 
         case 'shareList':
+          //If shareStatus is protected, and shareList is an array, then return combination of newMemberList and shareList, otherwise, return newMemberList, if shareStatus is not protected, then return null
+          const newShareList =
+            shareStatus === 'protected'
+              ? Array.isArray(shareList)
+                ? [...new Set(newMemberList.concat(shareList))]
+                : newMemberList
+              : null;
           foundVacation.shareList = newShareList;
+          break;
+
+        case 'endingTime':
+          //If endingTime < startingTime, then throw an error
+          req.body.endingTime < req.body.startingTime &&
+            _throw({
+              code: 400,
+              errors: [{ fields: 'endingTime', message: 'endingTime must be after startingTime' }],
+              message: 'invalid endingTime',
+            });
           break;
 
         default:
@@ -280,28 +216,20 @@ const vacationController = {
     await foundVacation.save();
 
     //Send to front
-    return res.status(200).json({ data: foundVacation, meta: '', message: 'update successfully' });
+    return res.status(201).json({ data: foundVacation, meta: '', message: 'update successfully' });
   }),
 
   delete: asyncWrapper(async (req, res) => {
     const { id } = req.params;
 
-    //Throw an error if cannot find Vacation
-    const foundVacation = await Vacations.findById(id);
-    !foundVacation &&
-      _throw({ code: 400, errors: [{ field: 'id', message: 'vacationId did not existed' }], message: 'invalid id' });
+    //Check whether user login is author of this vacation, if user is not author, then throw an error Forbidden
+    await checkAuthor({ type: 'vacation', vacationId: id, userId: req.userInfo._id });
 
-    //Throw an error if user is not author of vacation
-    const userId = req.userInfo._id.toString();
-    foundVacation.userId.toString() !== userId &&
-      _throw({
-        code: 400,
-        errors: [{ field: 'userId', message: 'userId is not author of this vacation' }],
-        message: 'invalid userId',
-      });
-
+    //Delete Vacation
     const deleteVacation = await Vacations.findByIdAndDelete(id);
-    return res.status(200).json({ data: deleteVacation, message: 'delete successfully' });
+
+    //Send to front
+    return res.status(202).json({ data: deleteVacation, message: 'delete successfully' });
   }),
 };
